change listener to dialogs position
