+++ app/src/main/res/layout/activity_maps.xml
@@ -24,7 +24,7 @@
         app:cardCornerRadius="5dp">
 
         <androidx.appcompat.widget.AppCompatEditText
-            android:id="@+id/search"
+            android:id="@+id/EditTextSearchValue"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="@android:color/white"
