+++ app/src/main/res/layout/confirm_order_dialog.xml
@@ -147,14 +147,13 @@
             android:id="@+id/textView6"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="60dp"
             android:layout_marginTop="11dp"
-            android:layout_marginEnd="85dp"
+            android:layout_marginEnd="100dp"
             android:text="Per Km"
             android:textSize="16sp"
             android:textStyle="bold"
             app:layout_constraintEnd_toStartOf="@+id/textView7"
-            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintStart_toStartOf="@+id/textViewSelectedInventory"
             app:layout_constraintTop_toBottomOf="@+id/imageView2" />
 
         <TextView
@@ -172,14 +171,13 @@
             android:id="@+id/textView8"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginStart="60dp"
             android:layout_marginTop="18dp"
-            android:layout_marginEnd="107dp"
+            android:layout_marginEnd="123dp"
             android:text="Total"
             android:textSize="19sp"
             android:textStyle="bold"
             app:layout_constraintEnd_toStartOf="@+id/finalPriceTextView"
-            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintStart_toStartOf="@+id/textViewSelectedInventory"
             app:layout_constraintTop_toBottomOf="@+id/textView6" />
 
         <TextView
